+++ tests/test_tchannel.py
@@ -36,12 +36,23 @@ import pytest
 from tornado import gen
 
 from tchannel.tornado.stream import InMemStream
-from tchannel import TChannel, Request, Response, schemes, errors
+from tchannel import TChannel, Request, Response, schemes, errors, thrift
 from tchannel.errors import AlreadyListeningError, TimeoutError
 from tchannel.event import EventHook
 from tchannel.response import TransportHeaders
 
 # TODO - need integration tests for timeout and retries, use testing.vcr
+
+
+@pytest.fixture
+def thrift_module(tmpdir, request):
+    thrift_file = tmpdir.join('service.thrift')
+    thrift_file.write('''
+        service Service {
+            bool healthy()
+        }
+    ''')
+    return thrift.load(str(thrift_file), request.node.name)
 
 
 @pytest.mark.call
@@ -515,5 +526,51 @@ def test_response_status_is_copied():
         arg2=b'\x00\x00',
         arg3=b'\x00',
     )
-
     assert 1 == response.status
+
+
+@pytest.mark.gen_test
+def test_per_request_caller_name_raw():
+    server = TChannel('server')
+    server.listen()
+
+    @server.raw.register('foo')
+    def handler(request):
+        assert request.transport.caller_name == 'bar'
+        return b'success'
+
+    client = TChannel('client', known_peers=[server.hostport])
+    res = yield client.raw('service', 'foo', b'', caller_name='bar')
+    assert res.body == b'success'
+
+
+@pytest.mark.gen_test
+def test_per_request_caller_name_json():
+    server = TChannel('server')
+    server.listen()
+
+    @server.json.register('foo')
+    def handler(request):
+        assert request.transport.caller_name == 'bar'
+        return {'success': True}
+
+    client = TChannel('client', known_peers=[server.hostport])
+    res = yield client.json('service', 'foo', {}, caller_name='bar')
+    assert res.body == {'success': True}
+
+
+@pytest.mark.gen_test
+def test_per_request_caller_name_thrift(thrift_module):
+    server = TChannel('server')
+    server.listen()
+
+    @server.thrift.register(thrift_module.Service)
+    def healthy(request):
+        assert request.transport.caller_name == 'bar'
+        return True
+
+    client = TChannel('client', known_peers=[server.hostport])
+    res = yield client.thrift(
+        thrift_module.Service.healthy(), caller_name='bar',
+    )
+    assert res.body is True